+++ src/presenter/points-presenter.js
@@ -8,8 +8,8 @@ import EditingPoint from '../view/point-edit.js';
 import NewTripPoint from './new-point-presenter.js';
 import PriceTripView from '../view/price-trip.js';
 import TripPoint from './point-presenter.js';
-import { RenderPosition, render, remove } from '../utils/rendering-utils.js';
-import { SortType, UpdateType, UserAction, FilterType, State } from '../utils/constants.js';
+import { render, remove } from '../utils/rendering-utils.js';
+import { SortType, UpdateType, UserAction, FilterType, State, RenderPosition } from '../utils/constants.js';
 import dayjs from 'dayjs';
 import { filter } from '../utils/filter.js';
 import Api from '../api.js';
